Allow command-line arguments to be grouped in the usage. (#8)
The @arg annotation has a "group" member that can be set.  Leaving it
blank (an empty string) will cause it to be part of the default group,
and listed at the top of the usage.  The groups are then displayed in
the usage in alphabetical order, with sections for required and optional
command-line arguments for each group.

Includes a fix for missing square brackets when outputting the default
and true/false values.

Optional arguments with an "empty" default (ex. Option with a default of
None, or Seq with a default of Seq.empty), display "Optional" instead of
nothing at the end of the command line description.